[config] move consensus_config into own file
purely logic move

Closes: #1722
Approved by: mimoo
